requirements.txt
pyqt5
numpy
tqdm
pillow
pytorch
torchvision
scikit-learn